+++ realm/realm-library/src/main/java/io/realm/BaseRealm.java
@@ -29,12 +29,12 @@ import java.util.concurrent.atomic.AtomicBoolean;
 import io.realm.exceptions.RealmFileException;
 import io.realm.exceptions.RealmMigrationNeededException;
 import io.realm.internal.CheckedRow;
+import io.realm.internal.ColumnInfo;
 import io.realm.internal.InvalidRow;
 import io.realm.internal.ObjectServerFacade;
 import io.realm.internal.RealmObjectProxy;
-import io.realm.internal.SharedRealm;
-import io.realm.internal.ColumnInfo;
 import io.realm.internal.Row;
+import io.realm.internal.SharedRealm;
 import io.realm.internal.Table;
 import io.realm.internal.UncheckedRow;
 import io.realm.internal.Util;
@@ -61,7 +61,7 @@ abstract class BaseRealm implements Closeable {
             "Changing Realm data can only be done from inside a transaction.";
     static final String LISTENER_NOT_ALLOWED_MESSAGE = "Listeners cannot be used on current thread.";
 
-    
+
     volatile static Context applicationContext;
 
     // Thread pool for all async operations (Query & transaction)
@@ -580,7 +580,8 @@ abstract class BaseRealm implements Closeable {
     /**
      * Migrates the Realm file defined by the given configuration using the provided migration block.
      *
-     * @param configuration configuration for the Realm that should be migrated.
+     * @param configuration configuration for the Realm that should be migrated. If this is a SyncConfiguration this
+     *                      method does nothing.
      * @param migration if set, this migration block will override what is set in {@link RealmConfiguration}.
      * @param callback callback for specific Realm type behaviors.
      * @param cause which triggers this migration.
@@ -589,8 +590,12 @@ abstract class BaseRealm implements Closeable {
     protected static void migrateRealm(final RealmConfiguration configuration, final RealmMigration migration,
                                        final MigrationCallback callback, final RealmMigrationNeededException cause)
             throws FileNotFoundException {
+
         if (configuration == null) {
             throw new IllegalArgumentException("RealmConfiguration must be provided");
+        }
+        if (configuration.isSyncConfiguration()) {
+            return;
         }
         if (migration == null && configuration.getMigration() == null) {
             throw new RealmMigrationNeededException(configuration.getPath(), "RealmMigration must be provided", cause);