+++ realm/realm-library/src/objectServer/java/io/realm/ObjectServer.java
@@ -19,17 +19,14 @@ package io.realm;
 import android.content.Context;
 import android.content.pm.PackageInfo;
 
-import io.realm.annotations.Beta;
 import io.realm.internal.Keep;
 
 /**
- * @Beta
  * Internal initializer class for the Object Server.
  * Use to keep the `SyncManager` free from Android dependencies
  */
 @SuppressWarnings("unused")
 @Keep
-@Beta
 class ObjectServer {
 
     public static void init(Context context) {
@@ -45,5 +42,10 @@ class ObjectServer {
         UserStore userStore = new RealmFileUserStore(context.getFilesDir().getPath());
 
         SyncManager.init(appId, userStore);
+
+        // init the "sync_manager.cpp" metadata Realm, this is also needed later, when re try
+        // to schedule a client reset. in realm-java#master this is already done, when initialising
+        // the RealmFileUserStore (not available now on releases)
+        SyncManager.nativeConfigureMetaDataSystem(context.getFilesDir().getPath());
     }
 }