+++ realm/realm-library/src/syncIntegrationTest/java/io/realm/objectserver/AuthTests.java
@@ -1,21 +1,22 @@
 package io.realm.objectserver;
 
-import android.support.test.InstrumentationRegistry;
 import android.support.test.runner.AndroidJUnit4;
 
-import org.junit.AfterClass;
-import org.junit.BeforeClass;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import io.realm.SyncCredentials;
 import io.realm.ErrorCode;
 import io.realm.ObjectServerError;
 import io.realm.Realm;
+import io.realm.SessionState;
+import io.realm.SyncConfiguration;
+import io.realm.SyncCredentials;
+import io.realm.SyncManager;
+import io.realm.SyncSession;
 import io.realm.SyncUser;
 import io.realm.objectserver.utils.Constants;
-import io.realm.objectserver.utils.HttpUtils;
+import io.realm.objectserver.utils.UserFactory;
 import io.realm.rule.RunInLooperThread;
 import io.realm.rule.RunTestInLooperThread;
 
@@ -23,20 +24,9 @@ import static junit.framework.Assert.assertEquals;
 import static junit.framework.Assert.fail;
 
 @RunWith(AndroidJUnit4.class)
-public class AuthTests {
+public class AuthTests extends BaseIntegrationTest {
     @Rule
     public RunInLooperThread looperThread = new RunInLooperThread();
-
-    @BeforeClass
-    public static void setUp () throws Exception {
-        Realm.init(InstrumentationRegistry.getContext());
-        HttpUtils.startSyncServer();
-    }
-
-    @AfterClass
-    public static void tearDown () throws Exception {
-        HttpUtils.stopSyncServer();
-    }
 
     @Test
     public void login_userNotExist() {
@@ -65,5 +55,71 @@ public class AuthTests {
                 looperThread.testComplete();
             }
         });
+    }
+
+    @Test
+    @RunTestInLooperThread
+    public void login_withAccessToken() {
+        SyncUser admin = UserFactory.createAdminUser(Constants.AUTH_URL);
+        SyncCredentials credentials = SyncCredentials.accessToken(admin.getAccessToken(), "custom-admin-user");
+        SyncUser.loginAsync(credentials, Constants.AUTH_URL, new SyncUser.Callback() {
+            @Override
+            public void onSuccess(SyncUser user) {
+                final SyncConfiguration config = new SyncConfiguration.Builder(user, Constants.SYNC_SERVER_URL)
+                        .errorHandler(new SyncSession.ErrorHandler() {
+                            @Override
+                            public void onError(SyncSession session, ObjectServerError error) {
+                                fail("Session failed: " + error);
+                            }
+                        })
+                        .build();
+
+                final Realm realm = Realm.getInstance(config);
+                looperThread.testRealms.add(realm);
+
+                // FIXME: Right now we have no Java API for detecting when a session is established
+                // So we optimistically assume it has been connected after 1 second.
+                looperThread.postRunnableDelayed(new Runnable() {
+                    @Override
+                    public void run() {
+                        assertEquals(SessionState.BOUND, SyncManager.getSession(config).getState());
+                        looperThread.testComplete();
+                    }
+                }, 1000);
+            }
+
+            @Override
+            public void onError(ObjectServerError error) {
+                fail("Login failed: " + error);
+            }
+        });
+    }
+
+    // The error handler throws an exception but it is ignored (but logged). That means, this test should not
+    // pass and not be stopped by an IllegalArgumentException.
+    @Test
+    @RunTestInLooperThread
+    public void loginAsync_errorHandlerThrows() {
+        SyncCredentials credentials = SyncCredentials.usernamePassword("IWantToHackYou", "GeneralPassword", false);
+        SyncUser.loginAsync(credentials, Constants.AUTH_URL, new SyncUser.Callback() {
+            @Override
+            public void onSuccess(SyncUser user) {
+                fail();
+            }
+
+            @Override
+            public void onError(ObjectServerError error) {
+                assertEquals(ErrorCode.INVALID_CREDENTIALS, error.getErrorCode());
+                throw new IllegalArgumentException("BOOM");
+            }
+        });
+
+        try {
+            Thread.sleep(2000);
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+            fail();
+        }
+        looperThread.testComplete();
     }
 }