+++ realm/realm-library/src/syncIntegrationTest/java/io/realm/objectserver/utils/UserFactory.java
@@ -16,25 +16,24 @@
 
 package io.realm.objectserver.utils;
 
-import java.net.URI;
-import java.net.URISyntaxException;
-
+import io.realm.SyncCredentials;
 import io.realm.SyncUser;
-import io.realm.objectserver.utils.Constants;
 
 // Must be in `io.realm.objectserver` to work around package protected methods.
 public class UserFactory {
-    // FIXME: Not working right now.
-    /*
-    public static User createDefaultUser(String SERVER_URL, String USER_TOKEN) {
-        try {
-            User user = User.createLocal();
 
-            user.addAccessToken(new URI(SERVER_URL), USER_TOKEN);
-            return user;
-        } catch (URISyntaxException e) {
-            throw new RuntimeException(e);
-        }
+    public static SyncUser createDefaultUser(String authUrl) {
+        return createUser(authUrl, "test-user");
     }
-    */
+
+    public static SyncUser createUser(String authUrl, String userIdentifier) {
+        SyncCredentials credentials = SyncCredentials.usernamePassword(userIdentifier, "myPassw0rd", true);
+        return SyncUser.login(credentials, authUrl);
+    }
+
+    public static SyncUser createAdminUser(String authUrl) {
+        // `admin` required as user identifier to be granted admin rights.
+        SyncCredentials credentials = SyncCredentials.custom("admin", "debug", null);
+        return SyncUser.login(credentials, authUrl);
+    }
 }